Move discovery options to genesis definition (#3113)
