+++ besu/src/test/java/org/hyperledger/besu/RunnerTest.java
@@ -16,7 +16,6 @@ package org.hyperledger.besu;
 
 import static java.util.Collections.emptySet;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hyperledger.besu.cli.config.EthNetworkConfig.DEV_NETWORK_ID;
 import static org.hyperledger.besu.cli.config.NetworkName.DEV;
 
 import org.hyperledger.besu.cli.config.EthNetworkConfig;
@@ -269,7 +268,7 @@ public final class RunnerTest {
       final EthNetworkConfig behindEthNetworkConfiguration =
           new EthNetworkConfig(
               EthNetworkConfig.jsonConfig(DEV),
-              DEV_NETWORK_ID,
+              DEV.getNetworkId(),
               Collections.singletonList(enode),
               null);
       runnerBehind =
