+++ config/src/main/resources/astor.json
@@ -4,6 +4,11 @@
     "chainId": 212,
     "ecip1049Block": 0,
     "keccak256": {
+    },
+    "discovery": {
+      "bootnodes": [
+        "enode://b638fc3dca6181ae97fac2ea0157e8330f5ac8a20c0d4c63aa6f98dcbac4e35b4e023f656757b58c1da7a7b2be9ffad9342e0f769b8cf0f5e35ff73116ff7dfd@3.16.171.213:30303"
+      ]
     }
   },
   "nonce": "0x42",
